requirements.txt
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP 客户端
httpx==0.27.2

# 数据验证
pydantic==2.9.2
pydantic-settings==2.6.0

# 配置管理
pyyaml==6.0.2

# 测试
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# 开发工具
black==24.10.0
flake8==7.1.1
mypy==1.13.0
